\name{BridgeDbR-package}
\alias{BridgeDbR-package}
\alias{BridgeDbR}
\docType{package}
\title{
BridgeDbR Package.
}
\description{
R package for using BridgeDb directly from R.
}
\author{
Christ Leemans
Anwesha Bohler
Denise Slenter
Lars Eijssen
Tooba Abbassi-Daloii

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ BridgeDB }
\keyword{ identifier mapping }
\keyword{ WikiPathways }
