\name{BridgeDbR-package}
\alias{BridgeDbR-package}
\alias{BridgeDbR}
\docType{package}
\title{
BridgeDBR Package.
}
\description{
R package for using BridgeDB directly from R.
}
\author{
Christ Leemans

Maintainer: Anwesha Bohler <anweshabohler@gmail.com>
}
\keyword{ package }
\keyword{ BridgeDB }
\keyword{ identifier mapping }
\keyword{ WikiPathways }
